% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryBee.R, R/summaryPlant.R
\docType{data}
\name{data}
\alias{data}
\title{Summary statistics for bees}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1000 rows and 26 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1000 rows and 26 columns.
}
\usage{
data

data
}
\arguments{
\item{input}{None}
}
\value{
Prints number of unique bee species, genuses, tribes, and families. Also returns a table of the number of observations for each species.

Prints number of unique plant species, genuses and families. Also returns a table of the number of observations for each species.
}
\description{
Function that returns summary of bee observations by species, genus, tribe, and family

Function that returns summary of plant species in the GloBI curated data.
}
\details{
Designed to work automatically on GloBI_Curated_sample data

Designed to work automatically on GloBI_Curated_sample data
}
\examples{
summaryBee()

summaryPlant()

}
\keyword{datasets}
