% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GloBI_Curated_sample.R
\docType{data}
\name{GloBI_Curated_sample}
\alias{GloBI_Curated_sample}
\title{GloBI_Curated_sample}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1000 rows and 26 columns.
}
\usage{
GloBI_Curated_sample
}
\description{
GloBI_Curated_sample
}
\keyword{datasets}
