% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryBee.R
\name{summaryBee}
\alias{summaryBee}
\title{Summary statistics for bees}
\usage{
summaryBee()
}
\arguments{
\item{input}{None}
}
\value{
Prints number of unique bee species, genuses, tribes, and families. Also returns a table of the number of observations for each species.
}
\description{
Function that returns summary of bee observations by species, genus, tribe, and family
}
\details{
Designed to work automatically on GloBI_Curated_sample data
}
\examples{
summaryBee()

}
