% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryData.R
\name{summaryData}
\alias{summaryData}
\title{summarizes variables and unique entries}
\usage{
summaryData(data = GloBI_Curated_sample, variable = NULL, print = FALSE)
}
\arguments{
\item{data}{GloBI data, defaults to GloBI_Curated_sample}

\item{variable}{Variable to color by, defaults to NULL}

\item{print}{Variable to print, defaults to TRUE}
}
\value{
Returns a list of variables in the GloBI df or if specified
}
\description{
Summarizes variables and unique entries
}
\details{
No variables are needed. Data variable can be provided for GloBI data other than the curated sample. A variable name can be provided to get the unique entries of the variable.
}
\examples{
summaryData()
summaryData(variable = "bee_genus", print = TRUE)

}
