% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryInteract.R
\name{summaryInteract}
\alias{summaryInteract}
\title{Summary statistics for interactions}
\usage{
summaryInteract(x)
}
\arguments{
\item{input}{Brief description of input. (Add as many params as needed)}
}
\value{
Brief description of what function returns/prints.
}
\description{
Function that returns [INSERT SUMMARY STATISTICS]
}
\details{
Specifics about functionality, inputs, etc.
}
\examples{
Put function call as example.

}
