% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryPlant.R
\name{summaryPlant}
\alias{summaryPlant}
\title{Summary statistics for plants}
\usage{
summaryPlant()
}
\arguments{
\item{input}{None}
}
\value{
Prints number of unique plant species, genuses and families. Also returns a table of the number of observations for each species.
}
\description{
Function that returns summary of plant species in the GloBI curated data.
}
\details{
Designed to work automatically on GloBI_Curated_sample data
}
\examples{
summaryPlant()

}
