% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visBeeSpecies.R
\name{visBeeSpecies}
\alias{visBeeSpecies}
\title{Distribution of Bee Species by Family in interactions}
\usage{
visBeeSpecies(family = NULL)
}
\arguments{
\item{family}{A character string specifying which bee family the user wants to plot.
This parameter is optional and is not necessary to run the function.}
}
\value{
If no input is provided, this function returns a bar plot of bee species that is color-coded by its corresponding bee family.
If a bee family is specified, it generates a bar plot of bee species that belong to the specified family.
}
\description{
This function generates a bar plot of bee species within the GloBI_Curated_sample dataset.
The user has the option to specify the bee family they wish to see the species of.
}
\details{
This function is used to explore the types of bee species within the GloBI_Curated_sample dataset.
Users are able to view counts of each bee species and how they vary by bee family.
}
\examples{
visBeeSpecies()
visBeeSpecies("Apidae")
}
