% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visInteractMap.R
\name{visInteractMap}
\alias{visInteractMap}
\title{Interactive map detailing locations of bee-plant interactions}
\usage{
visInteractMap(data = GloBI_Curated_sample, variable = NULL, filter1 = NULL)
}
\arguments{
\item{data}{GloBI data, defaults to GloBI_Curated_sample}

\item{variable}{variable to filter by, defaults to NULL}

\item{filter1}{filter to apply to variable, defaults to NULL}
}
\value{
a leaflet map of all or subset of GloBI dataset provided
}
\description{
Creates a leaflet map showing the locations of bee-plant interactions.
}
\details{
allows user to input any GloBI dataset for visualization. The leaflet style map allows for exploration of the bee-planet interaction observations.
}
\examples{
visInteractMap()
visInteractMap(variable = "bee_genus", filter = "Lasioglossum")

}
