% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visInteractType.R
\name{visInteractType}
\alias{visInteractType}
\title{Types and counts of bee-plant interactions}
\usage{
visInteractType()
}
\value{
This function returns a bar plot of the types of bee-plant interactions.
}
\description{
This function generates a bar plot of bee-plant interaction types within the GloBI_Curated_sample dataset.
}
\details{
This function allows a user to explore the distribution of bee-plant interaction types.
Users are able to view how bees and plants most commonly interact.
This function requires no input.
}
\examples{
visInteractType()

}
