% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anovaPop.R
\name{anovaPop}
\alias{anovaPop}
\title{ANOVA analysis of bee and plant population dynamics}
\usage{
anovaPop(group, level)
}
\arguments{
\item{group}{bee or plant, which type of endemicity is being tested}

\item{level}{significance level of ANOVA regression}

\item{data}{GloBI data, defaults to GloBI_Curated_sample}
}
\value{
Prints message saying if ANOVA is significant or not.
}
\description{
Function that returns if the relationship between interaction type and bee/plant endemicity is significant.
}
\details{
Runs two-way ANVOA test with endemicity as the independent variable and interaction type as the dependent variable.
}
\examples{
anovaPop("bee", 0.05)
anovaPop("plant", 0.1)

}
