% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryBee.R
\name{summaryBee}
\alias{summaryBee}
\title{Summary statistics for bees}
\usage{
summaryBee(data = GloBI_Curated_sample)
}
\arguments{
\item{data}{GloBI data, defaults to GloBI_Curated_sample}
}
\value{
Prints number of unique bee species, genuses, tribes, and families. Also returns a table of the number of observations for each species.
}
\description{
Function that returns summary of bee observations by species, genus, tribe, and family
}
\details{
allows user to input any GloBI dataset.
}
\examples{
summaryBee()

}
