% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryGeo.R
\name{summaryGeo}
\alias{summaryGeo}
\title{Summary statistics for geographical distributions}
\usage{
summaryGeo(data = GloBI_Curated_sample)
}
\arguments{
\item{data}{GloBI data, defaults to GloBI_Curated_sample}
}
\value{
prints summary statistics for geographic distribution
}
\description{
Function that returns min, Q1, median, mean, Q2, max of geographical coordinate columns
}
\details{
allows user to input any GloBI dataset.
}
\examples{
summaryGeo()

}
