% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryGeo.R
\name{summaryGeo}
\alias{summaryGeo}
\title{Summary statistics for geographical distributions}
\usage{
summaryGeo(x)
}
\arguments{
\item{x}{Brief description of input. (Add as many params as needed)}
}
\value{
Brief description of what function returns/prints.
}
\description{
Function that returns INSERT SUMMARY STATISTICS
}
\details{
Specifics about functionality, inputs, etc.
}
\examples{
Put function call as example.

}
