% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryInteract.R
\name{summaryInteract}
\alias{summaryInteract}
\title{Summary statistics for interactions}
\usage{
summaryInteract()
}
\arguments{
\item{input}{none}
}
\value{
Prints a tibble summarizing the number of each interaction type, and counts of bee and plant species involved in each type of interaction.
}
\description{
Function that returns table of observations and species numbers per interaction type.
}
\details{
Works automatically on GloBI_Curated_sample data set.
}
\examples{

summaryInteract()

}
