% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visInteractMap.R
\name{visInteractMap}
\alias{visInteractMap}
\title{Interactive map detailing locations of bee-plant interactions}
\usage{
visInteractMap(data = GloBI_Curated_sample, by = NULL, filter = NULL)
}
\arguments{
\item{data}{GloBI data, defaults to GloBI_Curated_sample}

\item{by}{variable to color by, defaults to NULL}

\item{filter}{filter to apply to variable, defaults to NULL}
}
\value{
Brief description of what function returns/prints.
}
\description{
Add brief description to function(s)
}
\details{
Specifics about functionality, inputs, etc.
}
\examples{
visInteractMap()

}
