% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visInteractType.R
\name{visInteractType}
\alias{visInteractType}
\title{Types and counts of bee-plant interactions}
\usage{
visInteractType()
}
\arguments{
\item{input}{Brief description of input. (Add as many params as needed)}
}
\value{
Brief description of what function returns/prints.
}
\description{
Add brief description to function(s)
}
\details{
Specifics about functionality, inputs, etc.
}
\examples{
Put function call as example.

}
